requirements.txt
flask[async]
flask_minify
pillow
aiohttp
async_lru
markdown
beautifulsoup4
gunicorn